Revert "Table.cs extraction from TableReservation"
This reverts commit ea665450b5310a11784d23d59e5c943c5a63648d.
